refactor: view package 생성 및 view역할의 class들 분리
